% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ssnac_example.R
\name{plot_ssnac_example}
\alias{plot_ssnac_example}
\title{Plot SSNAC example graph}
\usage{
plot_ssnac_example(data)
}
\arguments{
\item{data}{A tidygraph (tbl_graph) object}
}
\value{
ggplot2 plot
}
\description{
Plot SSNAC example graph
}
